requirements.txt
python-dotenv==0.20.0
pathlib==1.0.1